painful componentization of threejs scene, need to figure out dynamic sizing and clean up
